+++ packages/components/src/presentation-components/properties/inputs/NumericPropertyInput.tsx
@@ -6,11 +6,11 @@
 import { forwardRef, useEffect, useImperativeHandle, useMemo, useRef, useState } from "react";
 import { PropertyValueFormat } from "@itwin/appui-abstract";
 import { Input } from "@itwin/itwinui-react";
-import { getDecimalRoundingError } from "./Utils.js";
+import { applyNumericConstraints, getDecimalRoundingError, getMinMaxFromPropertyConstraints } from "./Utils.js";
 
 import type { PrimitiveValue, PropertyDescription, PropertyRecord } from "@itwin/appui-abstract";
 import type { PropertyEditorProps } from "@itwin/components-react";
-import type { PropertyValueConstraints, WithConstraints } from "../../common/ContentBuilder.js";
+import type { WithConstraints } from "../../common/ContentBuilder.js";
 import type { PropertyEditorAttributes } from "../editors/Common.js";
 
 /** @internal */
@@ -34,11 +34,9 @@ export const NumericPropertyInput = forwardRef<PropertyEditorAttributes, Numeric
     setInputValue(newVal);
   };
 
-  const { min, max } = property.constraints
-    ? getMinMaxFromPropertyConstraints(property.constraints)
-    : { min: undefined, max: undefined };
+  const { min, max } = getMinMaxFromPropertyConstraints(property.constraints);
   const commitInput = () => {
-    const formattedInputValue = applyConstraints(inputValue, min, max);
+    const formattedInputValue = applyConstraints({ inputAsString: inputValue, min, max });
     setInputValue(formattedInputValue);
     onCommit && onCommit({ propertyRecord, newValue: parsePrimitiveValue(formattedInputValue) });
   };
@@ -167,32 +165,13 @@ export const NumericInput = forwardRef<PropertyEditorAttributes, NumericInputPro
 );
 NumericInput.displayName = "NumericInput";
 
-function applyConstraints(inputAsNumber: string, min: number | undefined, max: number | undefined): string {
+function applyConstraints({ inputAsString, min, max }: { inputAsString: string; min?: number; max?: number }): string {
   if (min === undefined && max === undefined) {
-    return inputAsNumber;
+    return inputAsString;
   }
-
-  if (!isFinite(Number(inputAsNumber))) {
-    return inputAsNumber;
+  const inputAsNumber = Number(inputAsString);
+  if (!isFinite(inputAsNumber)) {
+    return inputAsString;
   }
-
-  let valAsNumber = Number(inputAsNumber);
-  if (min !== undefined) {
-    valAsNumber = Math.max(valAsNumber, min);
-  }
-  if (max !== undefined) {
-    valAsNumber = Math.min(valAsNumber, max);
-  }
-  return valAsNumber.toString();
-}
-
-function getMinMaxFromPropertyConstraints(constraints: PropertyValueConstraints): {
-  min: number | undefined;
-  max: number | undefined;
-} {
-  if ("minimumValue" in constraints) {
-    return { min: constraints.minimumValue, max: constraints.maximumValue };
-  }
-
-  return { min: undefined, max: undefined };
+  return applyNumericConstraints({ value: inputAsNumber, min, max }).toString();
 }